prepared module for building
