+++ package.json
@@ -35,6 +35,10 @@
     "./cron": {
       "import": "./dist/lib/cron/index.js",
       "types": "./dist/lib/cron/index.d.ts"
+    },
+    "./logging": {
+      "import": "./dist/lib/logging/index.js",
+      "types": "./dist/lib/logging/index.d.ts"
     }
   },
   "files": [
